Refactor balatro_set_core and server modules
- Moved classes and constants from balatro_set_core to balatro_set_classes for better organization.
- Updated imports in server.py to reflect the new structure.
- Simplified joker and consumable ability triggers in the buy and sell functions.
- Made Negative a buy/sell self trigger
- Improved game state management by ensuring proper initialization of board and discard counts.
- Added scoring cards to the scoring context for better tracking during gameplay.
- Fixed some jokers
