+++ .gitignore
@@ -1,3 +1,4 @@
 leaderboard.json
 .vscode/settings.json
 *.pyc
+balatro-saves.json
